all fluent APIs should return Argument<T> rather than Argument
