+++ src/System.CommandLine/Argument.cs
@@ -6,7 +6,6 @@ using System.CommandLine.Binding;
 using System.CommandLine.Parsing;
 using System.CommandLine.Completions;
 using System.Linq;
-using System.IO;
 
 namespace System.CommandLine
 {
@@ -198,108 +197,6 @@ namespace System.CommandLine
 
         /// <inheritdoc />
         string IValueDescriptor.ValueName => Name;
-
-        /// <summary>
-        /// Adds completions for the argument.
-        /// </summary>
-        /// <param name="completions">The completions to add.</param>
-        /// <returns>The configured argument.</returns>
-        public Argument AddCompletions(params string[] completions)
-        {
-            Completions.Add(completions);
-            return this;
-        }
-
-        /// <summary>
-        /// Adds completions for the argument.
-        /// </summary>
-        /// <param name="completionsDelegate">A function that will be called to provide completions.</param>
-        /// <returns>The option being extended.</returns>
-        public Argument AddCompletions(Func<CompletionContext, IEnumerable<string>> completionsDelegate)
-        {
-            Completions.Add(completionsDelegate);
-            return this;
-        }
-
-        /// <summary>
-        /// Adds completions for the argument.
-        /// </summary>
-        /// <param name="completionsDelegate">A function that will be called to provide completions.</param>
-        /// <returns>The configured argument.</returns>
-        public Argument AddCompletions(Func<CompletionContext, IEnumerable<CompletionItem>> completionsDelegate)
-        {
-            Completions.Add(completionsDelegate);
-            return this;
-        }
-
-        /// <summary>
-        /// Configures the argument to accept only the specified values, and to suggest them as command line completions.
-        /// </summary>
-        /// <param name="values">The values that are allowed for the argument.</param>
-        /// <returns>The configured argument.</returns>
-        public Argument AcceptOnlyFromAmong(params string[] values)
-        {
-            AllowedValues?.Clear();
-            AddAllowedValues(values);
-            Completions.Clear();
-            Completions.Add(values);
-
-            return this;
-        }
-
-        /// <summary>
-        /// Configures the argument to accept only values representing legal file paths.
-        /// </summary>
-        /// <returns>The configured argument.</returns>
-        public Argument AcceptLegalFilePathsOnly()
-        {
-            var invalidPathChars = Path.GetInvalidPathChars();
-
-            AddValidator(result =>
-            {
-                for (var i = 0; i < result.Tokens.Count; i++)
-                {
-                    var token = result.Tokens[i];
-
-                    // File class no longer check invalid character
-                    // https://blogs.msdn.microsoft.com/jeremykuhne/2018/03/09/custom-directory-enumeration-in-net-core-2-1/
-                    var invalidCharactersIndex = token.Value.IndexOfAny(invalidPathChars);
-
-                    if (invalidCharactersIndex >= 0)
-                    {
-                        result.ErrorMessage = result.LocalizationResources.InvalidCharactersInPath(token.Value[invalidCharactersIndex]);
-                    }
-                }
-            });
-
-            return this;
-        }
-
-        /// <summary>
-        /// Configures the argument to accept only values representing legal file names.
-        /// </summary>
-        /// <remarks>A parse error will result, for example, if file path separators are found in the parsed value.</remarks>
-        /// <returns>The configured argument.</returns>
-        public Argument AcceptLegalFileNamesOnly()
-        {
-            var invalidFileNameChars = Path.GetInvalidFileNameChars();
-
-            AddValidator(result =>
-            {
-                for (var i = 0; i < result.Tokens.Count; i++)
-                {
-                    var token = result.Tokens[i];
-                    var invalidCharactersIndex = token.Value.IndexOfAny(invalidFileNameChars);
-
-                    if (invalidCharactersIndex >= 0)
-                    {
-                        result.ErrorMessage = result.LocalizationResources.InvalidCharactersInFileName(token.Value[invalidCharactersIndex]);
-                    }
-                }
-            });
-
-            return this;
-        }
 
         /// <summary>
         /// Parses a command line string value using the argument.
